initial commit
Adds two crates:

One is the core crate, `feco3`, which has the core functionality.
The other is `feco3_python`, which is a python binding to the core crate.

The reason for this separation is so we can add other bindings,
such as wasm in the future.

The python crate is turned into a python package, `_feco3`
with maturin, and then the pure-python package `feco3` imports
it to use. The reason for this is so that if we need any
translation or shimming to happen in the python layer.