Build out the python API more
There are many layers of wrapping and nesting here,
it might be nice to be able to remove some of this boilerplate. But,
I really do like how the translation layer is quite simple, and then I
can implement more complex and pythonic logic in the
python layer quite easily.
Currenlty pyo3 only supports
type hints etc through
.pyi stub files,
so if I'm going to have to keep those in sync, I figured
it might be worth it to
just do even more in python.