Switch to generics everywhere instead of dyn
I was having trouble in
arrow::RecordBatchProcessor. I *know* that the multi_writer
only holds RecordBatchWriters, but the compiler doesn't.
So I was having to do some downcasting from
Box<dyn RecordWriter> to RecordBatchWriter, and that
was being really tricky.

This allows each container to better understand what it is
holding.

It actually makes some things more succinct.